refactor(bmn): split auction-candidates page into smaller components (#340)
