+++ frontend/src/app/bmn/auction-candidates/page.tsx
@@ -10,12 +10,9 @@ import { Checkbox } from "@/components/ui/checkbox";
 import { cn } from "@/lib/utils";
 import {
   ArrowLeft,
-  ArrowUp,
-  ArrowDown,
   Eye,
   FileText,
   Gavel,
-  GripVertical,
   Loader2,
   Package,
   Printer,
@@ -23,124 +20,22 @@ import {
 } from "lucide-react";
 import { toast } from "sonner";
 
-interface AuctionAsset {
-  id: string;
-  kode_barang: string;
-  nup: string;
-  nup_lama?: string | null;
-  nama_barang: string;
-  jenis_bmn?: string | null;
-  merk_tipe?: string | null;
-  kondisi: string;
-  nilai_perolehan: number;
-  nilai_buku?: number | null;
-  satuan?: string | null;
-  lokasi_ruang?: string | null;
-  lokasi_spesifik?: string | null;
-  tahun_perolehan?: number | null;
-  no_polisi?: string | null;
-}
-
-interface AssetResponse {
-  data: AuctionAsset[];
-  last_page: number;
-  total?: number;
-}
-
-interface AttachmentPage {
-  assets: AuctionAsset[];
-  startIndex: number;
-  showColumnNumbers: boolean;
-  includeMeta: boolean;
-  includeTotal: boolean;
-  includeSignature: boolean;
-}
-
-const formatRupiah = (value: number | null | undefined) =>
-  new Intl.NumberFormat("id-ID", {
-    style: "currency",
-    currency: "IDR",
-    minimumFractionDigits: 0,
-  }).format(value || 0);
-
-const formatPlainRupiah = (value: number | null | undefined) =>
-  new Intl.NumberFormat("id-ID", { maximumFractionDigits: 0 }).format(value || 0);
-
-const formatDateLong = (date = new Date()) =>
-  new Intl.DateTimeFormat("id-ID", {
-    day: "numeric",
-    month: "long",
-    year: "numeric",
-  }).format(date);
-
-const dayNames = ["Minggu", "Senin", "Selasa", "Rabu", "Kamis", "Jumat", "Sabtu"];
-const numberWords = [
-  "",
-  "Satu",
-  "Dua",
-  "Tiga",
-  "Empat",
-  "Lima",
-  "Enam",
-  "Tujuh",
-  "Delapan",
-  "Sembilan",
-  "Sepuluh",
-  "Sebelas",
-];
-
-function numberToWords(value: number): string {
-  if (value < 12) return numberWords[value];
-  if (value < 20) return `${numberToWords(value - 10)} Belas`;
-  if (value < 100) {
-    const tens = Math.floor(value / 10);
-    const rest = value % 10;
-    return `${numberToWords(tens)} Puluh${rest ? ` ${numberToWords(rest)}` : ""}`;
-  }
-  if (value < 200) return `Seratus${value > 100 ? ` ${numberToWords(value - 100)}` : ""}`;
-  if (value < 1000) {
-    const hundreds = Math.floor(value / 100);
-    const rest = value % 100;
-    return `${numberToWords(hundreds)} Ratus${rest ? ` ${numberToWords(rest)}` : ""}`;
-  }
-  if (value < 2000) return `Seribu${value > 1000 ? ` ${numberToWords(value - 1000)}` : ""}`;
-  const thousands = Math.floor(value / 1000);
-  const rest = value % 1000;
-  return `${numberToWords(thousands)} Ribu${rest ? ` ${numberToWords(rest)}` : ""}`;
-}
-
-function getSpelledDate(date = new Date()) {
-  const day = dayNames[date.getDay()];
-  const dateText = numberToWords(date.getDate());
-  const month = new Intl.DateTimeFormat("id-ID", { month: "long" }).format(date);
-  const yearText = numberToWords(date.getFullYear());
-  return { day, dateText, month, yearText };
-}
-
-function getBaNumberSuffix(date = new Date()) {
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  return `K.18/TU/KAP.06.01/B/${month}/${date.getFullYear()}`;
-}
-
-function getSkNumberSuffix(date = new Date()) {
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  return `K.18/TU/KAP.05/${month}/${date.getFullYear()}`;
-}
-
-function shortenLokasi(value?: string | null) {
-  if (!value) return "-";
-  return value
-    .replace("Kantor Balai KSDA Kalimantan Timur", "Kantor Balai")
-    .replace("Seksi KSDA Wilayah I (Berau)", "Seksi Wil. I Berau")
-    .replace("Seksi KSDA Wilayah II (Tenggarong)", "Seksi Wil. II Tenggarong")
-    .replace("Seksi KSDA Wilayah III (Balikpapan)", "Seksi Wil. III Balikpapan");
-}
+import type { AuctionAsset, AssetResponse } from "./_lib/auction-helpers";
+import {
+  formatRupiah,
+  shortenLokasi,
+  getBaNumberSuffix,
+  getSkNumberSuffix,
+} from "./_lib/auction-helpers";
+import { ReorderPanel } from "./_components/ReorderPanel";
+import { SummaryTile } from "./_components/SummaryTile";
+import { CorrectionDocument, handlePrintBa } from "./_components/BaKoreksiDocument";
+import { SkPenghentianDocument, handlePrintSk } from "./_components/SkPenghentianDocument";
 
 export default function BmnAuctionCandidatesPage() {
   const [searchTerm, setSearchTerm] = useState("");
   const [page, setPage] = useState(1);
   const [perPage, setPerPage] = useState(10);
-  // orderedIds keeps the user-defined order for the document
   const [orderedIds, setOrderedIds] = useState<string[]>([]);
   const [showDocument, setShowDocument] = useState(false);
   const [showSkDocument, setShowSkDocument] = useState(false);
@@ -169,11 +64,8 @@ export default function BmnAuctionCandidatesPage() {
   });
 
   const assets = useMemo(() => response?.data || [], [response?.data]);
-
-  // selectedIds derived from orderedIds for O(1) lookup
   const selectedIds = useMemo(() => new Set(orderedIds), [orderedIds]);
 
-  // ordered selected assets — this is what the document uses
   const assetMap = useMemo(() => {
     const map = new Map<string, AuctionAsset>();
     assets.forEach((a) => map.set(a.id, a));
@@ -202,7 +94,6 @@ export default function BmnAuctionCandidatesPage() {
       setShowDocument(false);
       return;
     }
-    // preserve existing order, append newly selected
     setOrderedIds((prev) => {
       const existing = new Set(prev);
       const newIds = assets.filter((a) => !existing.has(a.id)).map((a) => a.id);
@@ -277,204 +168,8 @@ export default function BmnAuctionCandidatesPage() {
     }, 50);
   };
 
-  const handlePrint = () => {
-    if (orderedSelectedAssets.length === 0) {
-      toast.error("Tidak ada aset terpilih untuk dicetak.");
-      return;
-    }
-    const printContent = document.getElementById("ba-koreksi-print-root");
-    if (!printContent) return;
-    const printWindow = window.open("", "_blank");
-    if (!printWindow) return;
-
-    printWindow.document.write(`
-      <html>
-        <head>
-          <title>BA Koreksi Kondisi BMN</title>
-          <style>
-            @page { size: A4; margin: 0; }
-            body {
-              margin: 0;
-              padding: 0;
-              background: white;
-              color: black;
-              font-family: 'Bookman Old Style', Georgia, serif;
-              font-size: 11pt;
-              line-height: 1.25;
-            }
-            .ba-page {
-              width: 210mm;
-              min-height: 297mm;
-              box-sizing: border-box;
-              margin: 0 auto;
-              padding: 5mm 20mm 14mm;
-              page-break-after: always;
-            }
-            .ba-page:last-child { page-break-after: auto; }
-            .ba-header { margin-top: -5mm; margin-left: -16mm; margin-right: -16mm; text-align: center; }
-            .ba-header img { width: 196mm !important; max-width: 196mm !important; height: auto !important; }
-            .ba-body { width: 166mm; margin-left: auto; margin-right: auto; text-align: justify; text-justify: inter-word; }
-            .ba-body p { text-align: justify; text-justify: inter-word; }
-            .ba-attachment { width: 166mm; margin-left: auto; margin-right: auto; padding-top: 2.5rem; }
-            .ba-title { margin-top: 0.75rem; text-align: center; font-weight: 700; }
-            .ba-title p { margin: 0; line-height: 1.2; }
-            .ba-text-block { margin-top: 1rem; }
-            .ba-text-block > * + * { margin-top: 1.25rem; }
-            table { border-collapse: collapse; }
-            .identity-table td { padding: 0.125rem 0; }
-            .identity-table .label-cell { width: 24mm; }
-            .identity-table .colon-cell { width: 6mm; text-align: center; }
-            .asset-table { width: 100%; border-collapse: collapse; text-align: center; font-size: 8.5pt; }
-            .asset-table th, .asset-table td { border: 1px solid #000; padding: 0.25rem; }
-            .asset-table td.text-left { text-align: left; }
-            .asset-table td.text-right { text-align: right; }
-            .attachment-meta { width: 109mm; margin-left: auto; text-align: left; }
-            .attachment-meta .meta-row { display: grid; grid-template-columns: 24mm 5mm minmax(0, 1fr); align-items: start; }
-            .attachment-meta .meta-label { white-space: nowrap; }
-            .attachment-meta .meta-colon { text-align: center; }
-            .attachment-meta .meta-value { min-width: 0; }
-            .attachment-meta .lampiran-value { display: inline-block; max-width: 80mm; }
-            .signature { width: 20rem; margin-left: auto; }
-            .attachment-signature { margin-top: 3rem; }
-            .signature p { margin: 0; padding: 0; line-height: 1.15; }
-            .ttd-placeholder { box-sizing: border-box; height: 112px; padding-top: 40px; padding-left: 1.35cm; color: #94a3b8; }
-          </style>
-        </head>
-        <body>${printContent.innerHTML}</body>
-      </html>
-    `);
-    printWindow.document.close();
-    printWindow.focus();
-    setTimeout(() => printWindow.print(), 500);
-  };
-
-  const handlePrintSk = () => {
-    if (orderedSelectedAssets.length === 0) {
-      toast.error("Tidak ada aset terpilih untuk dicetak.");
-      return;
-    }
-    const printContent = document.getElementById("sk-penghentian-print-root");
-    if (!printContent) return;
-    const printWindow = window.open("", "_blank");
-    if (!printWindow) return;
-
-    printWindow.document.write(`
-      <html>
-        <head>
-          <title>SK Penghentian Penggunaan BMN</title>
-          <style>
-            @page { size: A4; margin: 0 0 28mm 0; }
-            @page sk-main { size: A4; margin: 12mm 0 28mm 0; }
-            @page sk-main:first { size: A4; margin: 0 0 28mm 0; }
-            @page sk-attachment { size: A4; margin: 0; }
-            * { box-sizing: border-box; }
-            body {
-              margin: 0; padding: 0; background: white; color: black;
-              font-family: 'Bookman Old Style', Georgia, serif;
-              font-size: 11pt; line-height: 1.4;
-            }
-            p { margin: 0; padding: 0; }
-            .sk-page {
-              width: 210mm;
-              margin: 0 auto; padding: 5mm 20mm 0;
-            }
-            .sk-main-document { page: sk-main; }
-            .sk-attachment-document {
-              page: sk-attachment;
-              page-break-before: always;
-              break-before: page;
-              min-height: 297mm;
-              padding: 12mm 20mm 28mm !important;
-            }
-            .sk-page-ttd { padding-bottom: 0; }
-            article { margin: 0; }
-            .sk-page-break { page-break-after: always; break-after: always; }
-            /* KOP */
-            .sk-kop {
-              margin-top: -5mm; margin-left: -16mm; margin-right: -16mm;
-              margin-bottom: 6px; text-align: center;
-            }
-            .sk-kop img { width: 196mm !important; max-width: 196mm !important; height: auto !important; display: block; margin: 0 auto; }
-            /* Judul SK — halaman 1 */
-            .sk-title {
-              width: 166mm; margin-left: auto; margin-right: auto;
-              margin-top: 10px; text-align: center; font-weight: bold; line-height: 1.3;
-            }
-            .sk-title-nomor { font-weight: normal; }
-            .sk-title-tentang { margin-top: 10px; }
-            /* Sub-judul (DENGAN RAHMAT, KEPALA BALAI) */
-            .sk-subtitle {
-              width: 166mm; margin-left: auto; margin-right: auto;
-              margin-top: 16px;
-            }
-            .sk-subtitle > p { text-align: center; font-weight: bold; }
-            .sk-subtitle > p + p { margin-top: 6px; }
-            /* Tabel Menimbang/Mengingat/Memutuskan */
-            .sk-body { width: 166mm; margin-left: auto; margin-right: auto; }
-            table { border-collapse: collapse; width: 100%; }
-            td { vertical-align: top; padding: 0; }
-
-            /* Menimbang/Mengingat: parent boleh paginate, item anak tetap utuh */
-            .sk-field-section {
-              display: grid;
-              grid-template-columns: 28mm 8mm minmax(0, 1fr);
-              break-inside: auto !important;
-              page-break-inside: auto !important;
-            }
-            .sk-field-section + .sk-field-section { margin-top: 0.75rem; }
-            .sk-field-label, .sk-field-colon { padding: 0; }
-            .sk-field-colon { text-align: center; }
-            .sk-mengingat-list {
-              break-inside: auto !important;
-              page-break-inside: auto !important;
-            }
-            .sk-mengingat-item {
-              display: grid;
-              grid-template-columns: 6mm minmax(0, 1fr);
-              break-inside: avoid !important;
-              page-break-inside: avoid !important;
-              padding-top: 0.35rem;
-            }
-            .sk-mengingat-item:first-child { padding-top: 0; }
-            .sk-mengingat-text { text-align: justify; }
-            /* Halaman 2 */
-            .sk-page2-body { width: 166mm; margin-left: auto; margin-right: auto; padding-top: 16mm; }
-            .sk-memutuskan { text-align: center; font-weight: bold; margin-bottom: 12px; }
-            /* TTD block */
-            .sk-ttd { width: 20rem; margin-left: auto; margin-top: 3rem; }
-            .sk-ttd, .sk-ttd p { font-weight: normal !important; text-align: left !important; }
-            .sk-ttd p { margin: 0; padding: 0; line-height: 1.3; }
-            .sk-signature-name { font-weight: bold !important; }
-            .ttd-placeholder { height: 112px; padding-top: 40px; padding-left: 1.35cm; color: #94a3b8; font-weight: normal !important; text-align: left !important; }
-            /* Tembusan */
-            .sk-tembusan { margin-top: 2rem; }
-            .sk-tembusan, .sk-tembusan p { font-weight: normal !important; text-align: left !important; }
-            .sk-tembusan p { margin: 0; padding: 0; line-height: 1.5; }
-            /* Halaman 3 lampiran */
-            .sk-attachment-meta { width: 109mm; margin-left: auto; text-align: left; }
-            .meta-row { display: grid; grid-template-columns: 24mm 5mm minmax(0, 1fr); align-items: start; }
-            .meta-label { white-space: nowrap; }
-            .meta-colon { text-align: center; }
-            .sk-lampiran-title { text-align: center; font-weight: bold; line-height: 1.3; margin-top: 1.5rem; margin-bottom: 0.75rem; }
-            .sk-asset-table { width: 100%; border-collapse: collapse; text-align: center; font-size: 8.5pt; }
-            .sk-asset-table th, .sk-asset-table td { border: 1px solid #000; padding: 0.25rem; }
-            .sk-asset-table thead { display: table-header-group; }
-            .sk-asset-table tr { break-inside: avoid; page-break-inside: avoid; }
-            .sk-column-number-row th { font-weight: normal; }
-            .sk-asset-table td.text-left { text-align: left; }
-            .sk-asset-table td.text-right { text-align: right; }
-            .sk-lampiran-ttd { width: 20rem; margin-left: auto; margin-top: 1rem; break-inside: avoid; page-break-inside: avoid; }
-            .sk-lampiran-ttd p { margin: 0; padding: 0; line-height: 1.15; }
-            .sk-lampiran-ttd .ttd-placeholder { height: 86px !important; padding-top: 28px !important; }
-          </style>
-        </head>
-        <body>${printContent.innerHTML}</body>
-      </html>
-    `);
-    printWindow.document.close();
-    printWindow.focus();
-    setTimeout(() => printWindow.print(), 500);
-  };
+  const handlePrint = () => handlePrintBa(orderedSelectedAssets);
+  const handlePrintSkDoc = () => handlePrintSk(orderedSelectedAssets, skNumber);
 
   return (
     <div className="p-6 md:p-10 space-y-8 animate-in fade-in slide-in-from-bottom-4 duration-500">
@@ -625,7 +320,7 @@ export default function BmnAuctionCandidatesPage() {
               </Button>
             )}
             {showSkDocument && (
-              <Button size="sm" variant="outline" className="rounded-lg text-xs" onClick={handlePrintSk}>
+              <Button size="sm" variant="outline" className="rounded-lg text-xs" onClick={handlePrintSkDoc}>
                 <Printer className="mr-1 h-3.5 w-3.5" />
                 Cetak SK
               </Button>
@@ -736,67 +431,16 @@ export default function BmnAuctionCandidatesPage() {
         </div>
       </div>
 
-      {/* Reorder panel — visible whenever assets are selected */}
       {orderedIds.length > 0 && (
-        <div className="rounded-2xl border border-zinc-200 bg-white p-4 dark:border-zinc-800 dark:bg-zinc-900 print:hidden">
-          <div className="mb-3 flex items-center justify-between">
-            <div>
-              <h2 className="text-sm font-bold text-zinc-900 dark:text-white">Urutan Aset Terpilih</h2>
-              <p className="text-xs text-zinc-500 dark:text-zinc-400">
-                Drag atau gunakan tombol ↑↓ untuk mengatur nomor urut di dokumen.
-              </p>
-            </div>
-            <span className="rounded-lg bg-red-50 px-2.5 py-1 text-xs font-bold text-red-700 dark:bg-red-500/10 dark:text-red-400">
-              {orderedIds.length} aset
-            </span>
-          </div>
-          <ol className="space-y-1.5">
-            {orderedSelectedAssets.map((asset, index) => (
-              <li
-                key={asset.id}
-                draggable
-                onDragStart={() => handleDragStart(index)}
-                onDragEnter={() => handleDragEnter(index)}
-                onDragEnd={handleDragEnd}
-                onDragOver={(e) => e.preventDefault()}
-                className="flex cursor-grab items-center gap-3 rounded-xl border border-zinc-100 bg-zinc-50 px-3 py-2 transition active:cursor-grabbing active:opacity-60 dark:border-zinc-800 dark:bg-zinc-800/50"
-              >
-                <GripVertical className="h-4 w-4 shrink-0 text-zinc-400" />
-                <span className="flex h-6 w-6 shrink-0 items-center justify-center rounded-full bg-red-600 text-[10px] font-black text-white">
-                  {index + 1}
-                </span>
-                <div className="min-w-0 flex-1">
-                  <p className="truncate text-xs font-semibold text-zinc-900 dark:text-zinc-100">{asset.nama_barang}</p>
-                  <p className="font-mono text-[10px] text-zinc-400">
-                    {asset.kode_barang} · NUP {asset.nup}
-                    {asset.merk_tipe ? ` · ${asset.merk_tipe}` : ""}
-                    {asset.no_polisi ? ` · ${asset.no_polisi}` : ""}
-                  </p>
-                </div>
-                <div className="flex shrink-0 gap-1">
-                  <button
-                    type="button"
-                    onClick={() => moveUp(index)}
-                    disabled={index === 0}
-                    aria-label="Pindah ke atas"
-                    className="flex h-7 w-7 items-center justify-center rounded-lg border border-zinc-200 bg-white text-zinc-500 transition hover:bg-zinc-100 disabled:opacity-30 dark:border-zinc-700 dark:bg-zinc-900 dark:hover:bg-zinc-700"
-                  >
-                    <ArrowUp className="h-3.5 w-3.5" />
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() => moveDown(index)}
-                    disabled={index === orderedIds.length - 1}
-                    aria-label="Pindah ke bawah"
-                    className="flex h-7 w-7 items-center justify-center rounded-lg border border-zinc-200 bg-white text-zinc-500 transition hover:bg-zinc-100 disabled:opacity-30 dark:border-zinc-700 dark:bg-zinc-900 dark:hover:bg-zinc-700"
-                  >
-                    <ArrowDown className="h-3.5 w-3.5" />
-                  </button>
-                </div>
-              </li>
-            ))}
-          </ol>
-        </div>
+        <ReorderPanel
+          orderedIds={orderedIds}
+          orderedSelectedAssets={orderedSelectedAssets}
+          onDragStart={handleDragStart}
+          onDragEnter={handleDragEnter}
+          onDragEnd={handleDragEnd}
+          onMoveUp={moveUp}
+          onMoveDown={moveDown}
+        />
       )}
 
       {showDocument && orderedSelectedAssets.length > 0 && (
@@ -822,7 +466,7 @@ export default function BmnAuctionCandidatesPage() {
               <h2 className="text-lg font-bold text-zinc-900 dark:text-white">Preview SK Penghentian Penggunaan BMN</h2>
               <p className="text-sm text-zinc-500 dark:text-zinc-400">Keputusan Kepala Balai tentang penghentian penggunaan aset terpilih.</p>
             </div>
-            <Button className="rounded-xl gap-2 bg-amber-600 text-xs hover:bg-amber-500" onClick={handlePrintSk}>
+            <Button className="rounded-xl gap-2 bg-amber-600 text-xs hover:bg-amber-500" onClick={handlePrintSkDoc}>
               <Printer className="h-3.5 w-3.5" />
               Cetak / Save PDF
             </Button>
@@ -830,772 +474,6 @@ export default function BmnAuctionCandidatesPage() {
           <SkPenghentianDocument assets={orderedSelectedAssets} skNumber={skNumber} />
         </section>
       )}
-    </div>
-  );
-}
-
-function SummaryTile({ label, value, tone }: { label: string; value: string; tone: "red" | "emerald" | "zinc" }) {
-  const toneClass = {
-    red: "bg-red-50 text-red-700 ring-red-100 dark:bg-red-500/10 dark:text-red-400 dark:ring-red-500/20",
-    emerald: "bg-emerald-50 text-emerald-700 ring-emerald-100 dark:bg-emerald-500/10 dark:text-emerald-400 dark:ring-emerald-500/20",
-    zinc: "bg-zinc-50 text-zinc-700 ring-zinc-100 dark:bg-zinc-800 dark:text-zinc-300 dark:ring-zinc-700",
-  }[tone];
-
-  return (
-    <div className="rounded-2xl border border-zinc-200 bg-white p-4 dark:border-zinc-800 dark:bg-zinc-900">
-      <p className="text-[10px] font-bold uppercase tracking-wider text-zinc-400">{label}</p>
-      <p className={cn("mt-2 inline-flex rounded-xl px-3 py-1 text-lg font-black ring-1", toneClass)}>{value}</p>
-    </div>
-  );
-}
-
-function CorrectionDocument({ assets, baNumber }: { assets: AuctionAsset[]; baNumber: string }) {
-  const today = new Date();
-  const { day, dateText, month, yearText } = getSpelledDate(today);
-  const baNumberText = `BA.${baNumber.trim() || "____"}/${getBaNumberSuffix(today)}`;
-  const datePhrase = `${dateText} bulan ${month} tahun ${yearText}`;
-
-  return (
-    <div id="ba-koreksi-print-root" className="ba-print-root space-y-6">
-      <style jsx global>{`
-        @media print {
-          body * { visibility: hidden; }
-          .ba-print-root, .ba-print-root * { visibility: visible; }
-          .ba-print-root {
-            position: absolute;
-            left: 0;
-            top: 0;
-            width: 100%;
-            background: white;
-            color: black;
-            font-family: 'Bookman Old Style', Georgia, serif;
-            font-size: 11pt;
-            line-height: 1.25;
-            margin: 0;
-            padding: 0;
-          }
-          .ba-page {
-            width: 210mm;
-            min-height: 297mm;
-            margin: 0 auto;
-            padding: 5mm 20mm 14mm;
-            box-shadow: none !important;
-            page-break-after: always;
-          }
-          .ba-header { margin-top: -5mm; margin-left: -16mm; margin-right: -16mm; }
-          .ba-header img { max-width: 196mm !important; }
-          .ba-body { width: 166mm; margin-left: auto; margin-right: auto; text-align: justify; text-justify: inter-word; }
-          .ba-body p { text-align: justify; text-justify: inter-word; }
-          .ba-attachment { width: 166mm; margin-left: auto; margin-right: auto; }
-          .attachment-meta { width: 109mm; margin-left: auto; text-align: left; }
-          .attachment-meta .meta-row { display: grid; grid-template-columns: 24mm 5mm minmax(0, 1fr); align-items: start; }
-          .attachment-meta .meta-label { white-space: nowrap; }
-          .attachment-meta .meta-colon { text-align: center; }
-          .attachment-meta .meta-value { min-width: 0; }
-          .attachment-meta .lampiran-value { display: inline-block; max-width: 80mm; }
-          .ba-page:last-child { page-break-after: auto; }
-          .ba-no-print { display: none !important; }
-        }
-      `}</style>
-      <article className="ba-page mx-auto min-h-100 max-w-[210mm] bg-white px-24 py-9 text-black shadow-xl ring-1 ring-zinc-200" style={{ fontFamily: "'Bookman Old Style', Georgia, serif", fontSize: "11pt", lineHeight: "1.25" }}>
-        <DocumentHeader />
-        <div className="ba-title mt-1 text-center font-bold leading-tight">
-          <p className="m-0">BERITA ACARA</p>
-          <p className="m-0">KOREKSI PERUBAHAN KONDISI BARANG MILIK NEGARA</p>
-          <p className="m-0 font-normal">Nomor : {baNumberText}</p>
-        </div>
-        <div className="ba-body ba-text-block mx-auto mt-4 w-[166mm] space-y-5 text-justify">
-          <p>
-            Pada hari {day} tanggal {datePhrase}, bertempat di Kantor Balai Konservasi Sumber Daya Alam Kalimantan Timur, kami penanggungjawab Unit Penatausahaan Kuasa Pengguna Barang pada Balai Konservasi Sumber Daya Alam Kalimantan Timur :
-          </p>
-          <table className="identity-table">
-            <tbody>
-              <tr><td className="label-cell w-24 py-0.5">Nama</td><td className="colon-cell w-6">:</td><td>M. ARI WIBAWANTO, S.Hut., M.Sc.</td></tr>
-              <tr><td className="label-cell py-0.5">NIP</td><td className="colon-cell">:</td><td>19740514 199903 1 001</td></tr>
-              <tr><td className="label-cell py-0.5">Jabatan</td><td className="colon-cell">:</td><td>Kepala Balai Konservasi Sumber Daya Alam Kalimantan Timur</td></tr>
-            </tbody>
-          </table>
-          <p>
-            Menyatakan bahwa telah dilakukan koreksi perubahan kondisi dengan cara melakukan koreksi terhadap kondisi Barang Milik Negara
-            pada Kantor Balai Konservasi Sumber Daya Alam Kalimantan Timur pada tanggal {datePhrase} berdasarkan Penilaian Barang Milik Negara
-            dengan hasil (rincian terlampir).
-          </p>
-          <p>
-            Demikian Berita Acara ini dibuat sebagai bahan koreksi perubahan kondisi Barang Milik Negara Semester Satu tahun {yearText},
-            dan apabila dikemudian hari terdapat kekeliruan akan dilakukan perbaikan sebagaimana mestinya.
-          </p>
-        </div>
-        <div className="signature mt-20 ml-auto w-80">
-          <p className="m-0">Unit Penatausaha Kuasa Pengguna Barang</p>
-          <p className="m-0">Kepala Balai,</p>
-          <div className="ttd-placeholder mt-4 h-28 box-border pt-10 pl-[1.35cm] text-zinc-400">${"{ttd_pengirim}"}</div>
-          <p className="m-0">M. ARI WIBAWANTO, S.Hut., M.Sc.</p>
-          <p className="m-0">NIP. 19740514 199903 1 001</p>
-        </div>
-      </article>
-
-      <article className="ba-page mx-auto min-h-100 max-w-[210mm] bg-white px-24 py-12 text-black shadow-xl ring-1 ring-zinc-200" style={{ fontFamily: "'Bookman Old Style', Georgia, serif", fontSize: "11pt", lineHeight: "1.25" }}>
-        <div className="ba-attachment mx-auto w-[166mm] pt-10">
-          <div className="attachment-meta ml-auto w-[109mm]">
-            <div className="meta-row grid grid-cols-[24mm_5mm_minmax(0,1fr)]">
-              <span className="meta-label whitespace-nowrap">Lampiran</span>
-              <span className="meta-colon text-center">:</span>
-              <span className="meta-value min-w-0">
-                <span className="lampiran-value inline-block max-w-[80mm]">Berita Acara Koreksi Perubahan Kondisi BMN</span>
-              </span>
-            </div>
-            <div className="meta-row grid grid-cols-[24mm_5mm_minmax(0,1fr)]">
-              <span className="meta-label whitespace-nowrap">Nomor</span>
-              <span className="meta-colon text-center">:</span>
-              <span className="meta-value min-w-0 whitespace-nowrap">{baNumberText}</span>
-            </div>
-            <div className="meta-row grid grid-cols-[24mm_5mm_minmax(0,1fr)]">
-              <span className="meta-label whitespace-nowrap">Tanggal</span>
-              <span className="meta-colon text-center">:</span>
-              <span className="meta-value min-w-0">{formatDateLong(today)}</span>
-            </div>
-          </div>
-          <AssetConditionTable title="I. Sebelum" assets={assets} mode="before" />
-          <AssetConditionTable title="II. Sesudah" assets={assets} mode="after" />
-          <div className="signature attachment-signature mt-12 ml-auto w-80">
-            <p className="m-0">Kepala Balai,</p>
-            <div className="ttd-placeholder mt-4 h-28 box-border pt-10 pl-[1.35cm] text-zinc-400">${"{ttd_pengirim}"}</div>
-            <p className="m-0">M. ARI WIBAWANTO, S.Hut., M.Sc.</p>
-            <p className="m-0">NIP. 19740514 199903 1 001</p>
-          </div>
-        </div>
-      </article>
-    </div>
-  );
-}
-
-function DocumentHeader() {
-  return (
-    <div className="ba-header -mx-18 text-center">
-      {/* eslint-disable-next-line @next/next/no-img-element */}
-      <img src="/header-terbaru.png" alt="Kop Surat" className="mx-auto h-auto w-full max-w-[196mm]" />
-    </div>
-  );
-}
-
-function AssetConditionTable({ title, assets, mode }: { title: string; assets: AuctionAsset[]; mode: "before" | "after" }) {
-  return (
-    <div className="mt-6">
-      <p className="mb-2 text-[12px] font-semibold">{title}</p>
-      <table className="asset-table w-full border-collapse text-center text-[10px]">
-        <thead>
-          <tr>
-            <th rowSpan={2} className="border border-black px-1 py-1">No.</th>
-            <th rowSpan={2} className="border border-black px-1 py-1">Kode Barang</th>
-            <th rowSpan={2} className="border border-black px-1 py-1">NUP</th>
-            <th rowSpan={2} className="border border-black px-1 py-1">Nama Barang</th>
-            <th rowSpan={2} className="border border-black px-1 py-1">Satuan</th>
-            <th rowSpan={2} className="border border-black px-1 py-1">Nilai Perolehan (Rp)</th>
-            <th colSpan={3} className="border border-black px-1 py-1">Kondisi</th>
-          </tr>
-          <tr>
-            <th className="border border-black px-1 py-1">Baik</th>
-            <th className="border border-black px-1 py-1">Rusak Ringan</th>
-            <th className="border border-black px-1 py-1">Rusak Berat</th>
-          </tr>
-        </thead>
-        <tbody>
-          {assets.map((asset, index) => (
-            <tr key={`${mode}-${asset.id}`}>
-              <td className="border border-black px-1 py-1">{index + 1}.</td>
-              <td className="border border-black px-1 py-1">{asset.kode_barang}</td>
-              <td className="border border-black px-1 py-1">{asset.nup}</td>
-              <td className="border border-black px-1 py-1 text-left">{asset.nama_barang}</td>
-              <td className="border border-black px-1 py-1">{asset.satuan || "Unit"}</td>
-              <td className="border border-black px-1 py-1 text-right">{formatPlainRupiah(asset.nilai_perolehan)}</td>
-              <td className="border border-black px-1 py-1">0</td>
-              <td className="border border-black px-1 py-1">{mode === "before" ? 1 : 0}</td>
-              <td className="border border-black px-1 py-1">{mode === "after" ? 1 : 0}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-
-function SkPenghentianDocument({ assets, skNumber }: { assets: AuctionAsset[]; skNumber: string }) {
-  const today = new Date();
-  const skNumberText = `SK.${skNumber.trim() || "____"}/${getSkNumberSuffix(today)}`;
-  const totalNilai = assets.reduce((sum, a) => sum + (a.nilai_perolehan || 0), 0);
-  const attachmentPages = useMemo<AttachmentPage[]>(() => {
-    const firstPageLimit = 15;
-    const continuationPageLimit = 17;
-    const lastPageLimit = 10;
-    const pages: AttachmentPage[] = [];
-    let cursor = 0;
-
-    const pushPage = (pageAssets: AuctionAsset[], startIndex: number, includeMeta: boolean) => {
-      pages.push({
-        assets: pageAssets,
-        startIndex,
-        showColumnNumbers: !includeMeta,
-        includeMeta,
-        includeTotal: false,
-        includeSignature: false,
-      });
-    };
-
-    if (assets.length <= lastPageLimit) {
-      pushPage(assets, 0, true);
-    } else if (assets.length <= firstPageLimit) {
-      const firstChunkSize = assets.length - 1;
-      pushPage(assets.slice(0, firstChunkSize), 0, true);
-      pushPage(assets.slice(firstChunkSize), firstChunkSize, false);
-    } else if (assets.length <= firstPageLimit + lastPageLimit) {
-      pushPage(assets.slice(0, firstPageLimit), 0, true);
-      pushPage(assets.slice(firstPageLimit), firstPageLimit, false);
-    } else {
-      pushPage(assets.slice(0, firstPageLimit), 0, true);
-      const continuationChunks: AuctionAsset[][] = [];
-      let lastPageStart = assets.length;
-
-      for (let i = firstPageLimit; i < assets.length; i += continuationPageLimit) {
-        const remaining = assets.length - i;
-
-        if (remaining <= continuationPageLimit + lastPageLimit) {
-          const previousPageSize = Math.min(continuationPageLimit, Math.max(1, remaining - 1));
-          continuationChunks.push(assets.slice(i, i + previousPageSize));
-          lastPageStart = i + previousPageSize;
-          break;
-        }
-
-        continuationChunks.push(assets.slice(i, i + continuationPageLimit));
-      }
-
-      if (lastPageStart < assets.length) {
-        continuationChunks.push(assets.slice(lastPageStart));
-      }
-
-      cursor = firstPageLimit;
-      for (const chunk of continuationChunks) {
-        pushPage(chunk, cursor, false);
-        cursor += chunk.length;
-      }
-    }
-
-    if (pages.length === 0) {
-      pushPage([], 0, true);
-    }
-
-    const lastPage = pages[pages.length - 1];
-    lastPage.includeTotal = true;
-    lastPage.includeSignature = true;
-
-    return pages;
-  }, [assets]);
-
-  const renderAttachmentTable = (chunk: AuctionAsset[], startIndex: number, showColumnNumbers: boolean, includeTotal: boolean) => (
-    <table className="sk-asset-table mt-4 w-full border-collapse text-center" style={{ fontSize: "8.5pt" }}>
-      <thead>
-        {showColumnNumbers && (
-          <tr className="sk-column-number-row" data-sk-measure="column-number-row">
-            <th className="border border-black px-1 py-1">1</th>
-            <th className="border border-black px-1 py-1">2</th>
-            <th className="border border-black px-1 py-1">3</th>
-            <th className="border border-black px-1 py-1">4</th>
-            <th className="border border-black px-1 py-1">5</th>
-            <th className="border border-black px-1 py-1">6</th>
-            <th className="border border-black px-1 py-1">7</th>
-            <th className="border border-black px-1 py-1">8</th>
-            <th className="border border-black px-1 py-1">9</th>
-            <th className="border border-black px-1 py-1">10</th>
-          </tr>
-        )}
-        <tr data-sk-measure={showColumnNumbers ? "header-numbered" : "header-simple"}>
-          <th className="border border-black px-1 py-1">No</th>
-          <th className="border border-black px-1 py-1">Kode Barang</th>
-          <th className="border border-black px-1 py-1">NUP</th>
-          <th className="border border-black px-1 py-1">Nama Barang</th>
-          <th className="border border-black px-1 py-1">Merk / Type</th>
-          <th className="border border-black px-1 py-1">No Polisi</th>
-          <th className="border border-black px-1 py-1">Tahun Perolehan</th>
-          <th className="border border-black px-1 py-1">Nilai Perolehan (Rp)</th>
-          <th className="border border-black px-1 py-1">Kondisi</th>
-          <th className="border border-black px-1 py-1">Keterangan</th>
-        </tr>
-      </thead>
-      <tbody>
-        {chunk.map((asset, index) => (
-          <tr key={asset.id} data-sk-row-index={startIndex + index}>
-            <td className="border border-black px-1 py-1">{startIndex + index + 1}.</td>
-            <td className="border border-black px-1 py-1">{asset.kode_barang}</td>
-            <td className="border border-black px-1 py-1">{asset.nup}</td>
-            <td className="border border-black px-1 py-1 text-left">{asset.nama_barang}</td>
-            <td className="border border-black px-1 py-1">{asset.merk_tipe || "-"}</td>
-            <td className="border border-black px-1 py-1">{asset.no_polisi || "-"}</td>
-            <td className="border border-black px-1 py-1">{asset.tahun_perolehan || "-"}</td>
-            <td className="border border-black px-1 py-1 text-right">{formatPlainRupiah(asset.nilai_perolehan)}</td>
-            <td className="border border-black px-1 py-1">{asset.kondisi}</td>
-            <td className="border border-black px-1 py-1">Surat Lengkap</td>
-          </tr>
-        ))}
-        {includeTotal && (
-          <tr data-sk-measure="total-row">
-            <td colSpan={7} className="border border-black px-1 py-1 text-center font-bold">Jumlah</td>
-            <td className="border border-black px-1 py-1 text-right font-bold">{formatPlainRupiah(totalNilai)}</td>
-            <td colSpan={2} className="border border-black px-1 py-1"></td>
-          </tr>
-        )}
-      </tbody>
-    </table>
-  );
-
-  const mengingat = [
-    "Undang-Undang Republik Indonesia Nomor 17 Tahun 2003 tentang Keuangan Negara;",
-    "Undang-Undang Republik Indonesia Nomor 1 Tahun 2004 tentang Perbendaharaan Negara;",
-    "Peraturan Pemerintah Nomor 27 Tahun 2014 tentang Pengelolaan Barang Milik Negara/Daerah sebagaimana telah diubah dengan Peraturan Pemerintah Nomor 28 Tahun 2020;",
-    "Peraturan Presiden Nomor 175 Tahun 2024 tentang Kementerian Kehutanan;",
-    "Peraturan Menteri Keuangan Nomor 4/PMK.06/2015 tentang Pendelegasian Kewenangan dan Tanggung Jawab Tertentu Dari Pengelola Barang kepada Pengguna Barang;",
-    "Peraturan Menteri Keuangan Nomor 83/PMK.06/2016 tentang Tata Cara Pelaksanaan Pemusnahan dan Penghapusan Barang Milik Negara;",
-    "Peraturan Menteri Keuangan Nomor 111/PMK.06/2016 tentang Tata Cara Pelaksanaan Pemindahtanganan Barang Milik Negara sebagaimana telah diubah dengan Peraturan Menteri Keuangan Nomor 165/PMK.06/2021;",
-    "Peraturan Menteri Keuangan Nomor 181/PMK.06/2016 tentang Penatausahaan Barang Milik Negara;",
-    "Peraturan Menteri Lingkungan Hidup dan Kehutanan Nomor P.11/MENLHK/SETJEN/KAP.3/4/2018 tentang Tata Cara Pelaksanaan Pemindahtanganan Barang Milik Negara Lingkup Kementerian Lingkungan Hidup dan Kehutanan.",
-  ];
-
-  const pageStyle: React.CSSProperties = {
-    fontFamily: "'Bookman Old Style', Georgia, serif",
-    fontSize: "11pt",
-    lineHeight: "1.4",
-  };
-
-  const renderAttachmentMetaTitle = (measure = false) => (
-    <div data-sk-measure={measure ? "attachment-meta-title" : undefined}>
-      <div className="sk-attachment-meta ml-auto w-[109mm]">
-        <div className="meta-row grid grid-cols-[24mm_5mm_minmax(0,1fr)]">
-          <span className="meta-label whitespace-nowrap">Lampiran</span>
-          <span className="meta-colon text-center">:</span>
-          <span className="meta-value min-w-0">Keputusan Kepala Balai KSDA KALTIM</span>
-        </div>
-        <div className="meta-row grid grid-cols-[24mm_5mm_minmax(0,1fr)]">
-          <span className="meta-label whitespace-nowrap">Nomor</span>
-          <span className="meta-colon text-center">:</span>
-          <span className="meta-value min-w-0 whitespace-nowrap">{skNumberText}</span>
-        </div>
-        <div className="meta-row grid grid-cols-[24mm_5mm_minmax(0,1fr)]">
-          <span className="meta-label whitespace-nowrap">Tanggal</span>
-          <span className="meta-colon text-center">:</span>
-          <span className="meta-value min-w-0">{formatDateLong(today)}</span>
-        </div>
-      </div>
-
-      <p className="sk-lampiran-title mt-6 text-center font-bold leading-snug">
-        DAFTAR PENGHENTIAN PENGGUNAAN BARANG MILIK NEGARA<br />
-        PADA BALAI KONSERVASI SUMBER DAYA ALAM KALIMANTAN TIMUR
-      </p>
-    </div>
-  );
-
-  const renderAttachmentSignature = (measure = false) => (
-    <div className="sk-lampiran-ttd signature mt-10 ml-auto w-80" data-sk-measure={measure ? "signature" : undefined}>
-      <p className="m-0">Kepala Balai,</p>
-      <div className="ttd-placeholder mt-4 h-28 box-border pt-10 pl-[1.35cm] text-zinc-400">${"{ttd_pengirim}"}</div>
-      <p className="m-0 font-bold">M. ARI WIBAWANTO, S.Hut., M.Sc.</p>
-      <p className="m-0">NIP. 19740514 199903 1 001</p>
-    </div>
-  );
-
-  return (
-    <div id="sk-penghentian-print-root" className="sk-print-root">
-      <style jsx global>{`
-        @media print {
-          @page { size: A4; margin: 0 0 28mm 0; }
-          @page sk-main { size: A4; margin: 12mm 0 28mm 0; }
-          @page sk-main:first { size: A4; margin: 0 0 28mm 0; }
-          @page sk-attachment { size: A4; margin: 0; }
-          body * { visibility: hidden; }
-          .sk-print-root, .sk-print-root * { visibility: visible; }
-          .sk-print-root {
-            position: absolute; left: 0; top: 0; width: 100%;
-            background: white; color: black;
-            font-family: 'Bookman Old Style', Georgia, serif;
-            font-size: 11pt; line-height: 1.4; margin: 0; padding: 0;
-          }
-          .sk-page {
-            width: 210mm; margin: 0 auto;
-            padding: 5mm 20mm 0; box-shadow: none !important;
-          }
-          .sk-main-document { page: sk-main; }
-          .sk-attachment-document {
-            page: sk-attachment;
-            page-break-before: always;
-            break-before: page;
-            min-height: 297mm;
-            padding: 12mm 20mm 28mm !important;
-          }
-          .sk-page-break { page-break-after: always; break-after: always; }
-          .sk-kop { margin-top: -5mm; margin-left: -16mm; margin-right: -16mm; margin-bottom: 4px; text-align: center; }
-          .sk-kop img { width: 196mm !important; max-width: 196mm !important; height: auto !important; display: block; margin: 0 auto; }
-          .sk-body { width: 166mm; margin-left: auto; margin-right: auto; }
-
-          .sk-mengingat-table { break-inside: auto !important; page-break-inside: auto !important; }
-          .sk-mengingat-row { break-inside: avoid !important; page-break-inside: avoid !important; }
-          .sk-field-section { display: grid; grid-template-columns: 28mm 8mm minmax(0, 1fr); break-inside: auto !important; page-break-inside: auto !important; }
-          .sk-field-section + .sk-field-section { margin-top: 0.75rem; }
-          .sk-field-colon { text-align: center; }
-          .sk-mengingat-list { break-inside: auto !important; page-break-inside: auto !important; }
-          .sk-mengingat-item { display: grid; grid-template-columns: 6mm minmax(0, 1fr); break-inside: avoid !important; page-break-inside: avoid !important; padding-top: 0.35rem; }
-          .sk-mengingat-item:first-child { padding-top: 0; }
-          .sk-mengingat-text { text-align: justify; }
-          .sk-no-print { display: none !important; }
-          .sk-asset-table { width: 100%; border-collapse: collapse; text-align: center; font-size: 8.5pt; }
-          .sk-asset-table th, .sk-asset-table td { border: 1px solid #000; padding: 0.25rem; }
-          .sk-asset-table thead { display: table-header-group; }
-          .sk-asset-table tr { break-inside: avoid; page-break-inside: avoid; }
-          .sk-column-number-row th { font-weight: normal; }
-          .sk-asset-table td.text-left { text-align: left; }
-          .sk-asset-table td.text-right { text-align: right; }
-          .sk-lampiran-ttd { margin-top: 1rem !important; break-inside: avoid; page-break-inside: avoid; }
-          .sk-lampiran-ttd p { line-height: 1.15 !important; }
-          .sk-attachment-meta .meta-row { display: grid; grid-template-columns: 24mm 5mm minmax(0, 1fr); align-items: start; }
-          .sk-attachment-meta .meta-label { white-space: nowrap; }
-          .sk-attachment-meta .meta-colon { text-align: center; }
-          .sk-ttd, .sk-ttd p { font-weight: normal !important; text-align: left !important; }
-          .signature p { margin: 0; padding: 0; line-height: 1.15; }
-          .sk-signature-name { font-weight: bold !important; }
-          .ttd-placeholder { box-sizing: border-box; height: 112px; padding-top: 40px; padding-left: 1.35cm; color: #94a3b8; font-weight: normal !important; text-align: left !important; }
-          .sk-lampiran-ttd .ttd-placeholder { height: 86px !important; padding-top: 28px !important; }
-          .sk-tembusan, .sk-tembusan p { font-weight: normal !important; text-align: left !important; }
-        }
-        .sk-print-root .sk-attachment-document {
-          min-height: 297mm;
-          padding: 12mm 20mm 28mm !important;
-        }
-        .sk-print-root .sk-page {
-          width: 210mm;
-          margin-left: auto;
-          margin-right: auto;
-          box-sizing: border-box;
-        }
-        .sk-print-root .sk-main-document {
-          padding: 5mm 20mm 0 !important;
-        }
-        .sk-print-root .sk-kop {
-          margin-top: -5mm;
-          margin-left: -16mm;
-          margin-right: -16mm;
-          margin-bottom: 4px;
-          text-align: center;
-        }
-        .sk-print-root .sk-kop img {
-          width: 196mm !important;
-          max-width: 196mm !important;
-          height: auto !important;
-          display: block;
-          margin: 0 auto;
-        }
-        .sk-print-root .sk-title,
-        .sk-print-root .sk-subtitle,
-        .sk-print-root .sk-body {
-          width: 166mm;
-          margin-left: auto;
-          margin-right: auto;
-        }
-        .sk-print-root .sk-title {
-          margin-top: 10px;
-          text-align: center;
-          font-weight: bold;
-          line-height: 1.3;
-        }
-        .sk-print-root .sk-title-nomor {
-          font-weight: normal;
-        }
-        .sk-print-root .sk-title-tentang {
-          margin-top: 10px;
-        }
-        .sk-print-root .sk-subtitle {
-          margin-top: 16px;
-        }
-        .sk-print-root .sk-subtitle > p {
-          text-align: center;
-          font-weight: bold;
-        }
-        .sk-print-root .sk-subtitle > p + p {
-          margin-top: 6px;
-        }
-        .sk-print-root p {
-          margin: 0;
-          padding: 0;
-        }
-        .sk-print-root table {
-          border-collapse: collapse;
-          width: 100%;
-        }
-        .sk-print-root td {
-          vertical-align: top;
-          padding: 0;
-        }
-        .sk-print-root .sk-field-section {
-          display: grid;
-          grid-template-columns: 28mm 8mm minmax(0, 1fr);
-          break-inside: auto;
-          page-break-inside: auto;
-        }
-        .sk-print-root .sk-field-section + .sk-field-section {
-          margin-top: 0.75rem;
-        }
-        .sk-print-root .sk-field-colon {
-          text-align: center;
-        }
-        .sk-print-root .sk-mengingat-list {
-          break-inside: auto;
-          page-break-inside: auto;
-        }
-        .sk-print-root .sk-mengingat-item {
-          display: grid;
-          grid-template-columns: 6mm minmax(0, 1fr);
-          break-inside: avoid;
-          page-break-inside: avoid;
-          padding-top: 0.35rem;
-        }
-        .sk-print-root .sk-mengingat-item:first-child {
-          padding-top: 0;
-        }
-        .sk-print-root .sk-mengingat-text {
-          text-align: justify;
-        }
-        .sk-print-root .sk-memutuskan {
-          text-align: center;
-          font-weight: bold;
-          margin-bottom: 12px;
-        }
-        .sk-print-root .sk-ttd {
-          width: 20rem;
-          margin-left: auto;
-          margin-top: 3rem;
-        }
-        .sk-print-root .sk-ttd,
-        .sk-print-root .sk-ttd p,
-        .sk-print-root .sk-tembusan,
-        .sk-print-root .sk-tembusan p {
-          font-weight: normal !important;
-          text-align: left !important;
-        }
-        .sk-print-root .sk-ttd p,
-        .sk-print-root .sk-tembusan p {
-          margin: 0;
-          padding: 0;
-        }
-        .sk-print-root .sk-signature-name {
-          font-weight: bold !important;
-        }
-        .sk-print-root .ttd-placeholder {
-          box-sizing: border-box;
-          height: 112px;
-          padding-top: 40px;
-          padding-left: 1.35cm;
-          color: #94a3b8;
-          font-weight: normal !important;
-          text-align: left !important;
-        }
-        .sk-print-root .sk-tembusan {
-          margin-top: 2rem;
-        }
-        .sk-print-root .sk-tembusan p {
-          line-height: 1.5;
-        }
-        .sk-print-root .sk-attachment-meta {
-          width: 109mm;
-          margin-left: auto;
-          text-align: left;
-        }
-        .sk-print-root .meta-row {
-          display: grid;
-          grid-template-columns: 24mm 5mm minmax(0, 1fr);
-          align-items: start;
-        }
-        .sk-print-root .meta-label {
-          white-space: nowrap;
-        }
-        .sk-print-root .meta-colon {
-          text-align: center;
-        }
-        .sk-print-root .sk-lampiran-title {
-          text-align: center;
-          font-weight: bold;
-          line-height: 1.3;
-          margin-top: 1.5rem;
-          margin-bottom: 0.75rem;
-        }
-        .sk-print-root .sk-asset-table {
-          width: 100%;
-          border-collapse: collapse;
-          text-align: center;
-          font-size: 8.5pt;
-        }
-        .sk-print-root .sk-asset-table th,
-        .sk-print-root .sk-asset-table td {
-          border: 1px solid #000;
-          padding: 0.25rem;
-        }
-        .sk-print-root .sk-asset-table tr {
-          break-inside: avoid;
-          page-break-inside: avoid;
-        }
-        .sk-print-root .sk-column-number-row th {
-          font-weight: normal;
-        }
-        .sk-print-root .sk-asset-table td.text-left {
-          text-align: left;
-        }
-        .sk-print-root .sk-asset-table td.text-right {
-          text-align: right;
-        }
-        .sk-print-root .sk-lampiran-ttd {
-          width: 20rem;
-          margin-left: auto;
-          margin-top: 1rem !important;
-          break-inside: avoid;
-          page-break-inside: avoid;
-        }
-        .sk-print-root .sk-lampiran-ttd p {
-          margin: 0;
-          padding: 0;
-          line-height: 1.15 !important;
-        }
-        .sk-print-root .sk-lampiran-ttd .ttd-placeholder {
-          height: 86px !important;
-          padding-top: 28px !important;
-        }
-      `}</style>
-
-      {/* ── HALAMAN 1+2: KOP + Judul + Menimbang + Mengingat + MEMUTUSKAN + TTD + Tembusan ── */}
-      <article
-        className="sk-page sk-page-ttd sk-main-document mx-auto max-w-[210mm] bg-white px-24 py-9 text-black shadow-xl ring-1 ring-zinc-200"
-        style={pageStyle}
-      >
-        <div className="sk-kop" style={{ marginTop: "-5mm", marginLeft: "-16mm", marginRight: "-16mm", marginBottom: "4px", textAlign: "center" }}>
-          {/* eslint-disable-next-line @next/next/no-img-element */}
-          <img src="/header-new.png" alt="Kop Surat" style={{ width: "196mm", maxWidth: "196mm", height: "auto", display: "block", margin: "0 auto" }} />
-        </div>
-
-        <div className="sk-title mx-auto mt-3 w-[166mm] text-center font-bold leading-snug">
-          <p className="m-0">KEPUTUSAN KEPALA BALAI</p>
-          <p className="m-0">KONSERVASI SUMBER DAYA ALAM KALIMANTAN TIMUR</p>
-          <p className="sk-title-nomor m-0 font-normal">Nomor : {skNumberText}</p>
-          <p className="sk-title-tentang m-0 mt-4">TENTANG</p>
-          <p className="m-0">PENGHENTIAN PENGGUNAAN BARANG MILIK NEGARA</p>
-          <p className="m-0">PADA BALAI KONSERVASI SUMBER DAYA ALAM KALIMANTAN TIMUR,</p>
-        </div>
-
-        <div className="sk-subtitle sk-body mx-auto mt-5 w-[166mm]">
-          <p className="text-center font-bold">DENGAN RAHMAT TUHAN YANG MAHA ESA</p>
-          <p className="mt-2 text-center font-bold">
-            KEPALA BALAI<br />
-            KONSERVASI SUMBER DAYA ALAM KALIMANTAN TIMUR,
-          </p>
-
-          {/* Menimbang + Mengingat */}
-          <div className="mt-4">
-            <div className="sk-field-section">
-              <div className="sk-field-label">Menimbang</div>
-              <div className="sk-field-colon">:</div>
-              <div className="sk-mengingat-list">
-                <div className="sk-mengingat-item">
-                  <div>a.</div>
-                  <div className="sk-mengingat-text">
-                    bahwa terdapat Barang Milik Negara pada Balai Konservasi Sumber Daya Alam Kalimantan Timur berupa Alat Angkutan Bermotor dalam keadaan rusak berat dan tidak ekonomis lagi untuk digunakan;
-                  </div>
-                </div>
-                <div className="sk-mengingat-item" style={{ paddingTop: "0.5rem" }}>
-                  <div>b.</div>
-                  <div className="sk-mengingat-text">
-                    bahwa sehubungan dengan hal tersebut diatas, dipandang perlu untuk menerbitkan Keputusan Kepala Balai Konservasi Sumber Daya Alam Kalimantan Timur tentang Penghentian Penggunaan Barang Milik Negara.
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            <div className="sk-field-section">
-              <div className="sk-field-label">Mengingat</div>
-              <div className="sk-field-colon">:</div>
-              <div className="sk-mengingat-list">
-                {mengingat.map((item, i) => (
-                  <div className="sk-mengingat-item" key={i}>
-                    <div>{i + 1}.</div>
-                    <div className="sk-mengingat-text">{item}</div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-
-          {/* MEMUTUSKAN */}
-          <p className="sk-memutuskan text-center font-bold" style={{ marginTop: "1.5rem" }}>MEMUTUSKAN</p>
-
-          <table className="sk-mengingat-table mt-4 w-full" style={{ borderCollapse: "collapse" }}>
-            <tbody>
-              <tr className="sk-mengingat-row">
-                <td style={{ width: "28mm", verticalAlign: "top" }}>Menetapkan</td>
-                <td style={{ width: "8mm", textAlign: "center", verticalAlign: "top" }}>:</td>
-                <td style={{ verticalAlign: "top", textTransform: "uppercase", textAlign: "justify" }}>
-                  KEPUTUSAN KEPALA BALAI KONSERVASI SUMBER DAYA ALAM KALIMANTAN TIMUR TENTANG PENGHENTIAN PENGGUNAAN BARANG MILIK NEGARA LINGKUP BALAI KONSERVASI SUMBER DAYA ALAM KALIMANTAN TIMUR.
-                </td>
-              </tr>
-              <tr className="sk-mengingat-row">
-                <td style={{ width: "28mm", verticalAlign: "top", paddingTop: "1rem" }}>KESATU</td>
-                <td style={{ width: "8mm", textAlign: "center", verticalAlign: "top", paddingTop: "1rem" }}>:</td>
-                <td style={{ verticalAlign: "top", paddingTop: "1rem", textAlign: "justify" }}>
-                  Menghentikan penggunaan Barang Milik Negara berupa Alat Angkutan Bermotor dalam kondisi rusak berat pada Balai Konservasi Sumber Daya Alam Kalimantan Timur tersebut sebagaimana tercantum dalam lampiran keputusan ini.
-                </td>
-              </tr>
-              <tr className="sk-mengingat-row">
-                <td style={{ width: "28mm", verticalAlign: "top", paddingTop: "1rem" }}>KEDUA</td>
-                <td style={{ width: "8mm", textAlign: "center", verticalAlign: "top", paddingTop: "1rem" }}>:</td>
-                <td style={{ verticalAlign: "top", paddingTop: "1rem", textAlign: "justify" }}>
-                  Menghentikan biaya pemeliharaan Alat Angkutan Bermotor tersebut sejak dikeluarkan keputusan ini, untuk dilanjutkan pada proses penghapusan.
-                </td>
-              </tr>
-              <tr className="sk-mengingat-row">
-                <td style={{ width: "28mm", verticalAlign: "top", paddingTop: "1rem" }}>KETIGA</td>
-                <td style={{ width: "8mm", textAlign: "center", verticalAlign: "top", paddingTop: "1rem" }}>:</td>
-                <td style={{ verticalAlign: "top", paddingTop: "1rem", textAlign: "justify" }}>
-                  Keputusan ini mulai berlaku sejak tanggal ditetapkan.
-                </td>
-              </tr>
-            </tbody>
-          </table>
-
-          {/* TTD */}
-          <div className="sk-ttd signature mt-12 ml-auto w-80">
-            <p className="m-0">Ditetapkan di : Samarinda</p>
-            <p className="m-0">Pada tanggal : {formatDateLong(today)}</p>
-            <p className="m-0 mt-3">Kepala Balai,</p>
-            <div className="ttd-placeholder mt-4 h-28 box-border pt-10 pl-[1.35cm] text-zinc-400">${"{ttd_pengirim}"}</div>
-            <p className="sk-signature-name m-0 font-bold">M. ARI WIBAWANTO, S.Hut., M.Sc.</p>
-            <p className="m-0">NIP. 19740514 199903 1 001</p>
-          </div>
-
-          {/* Tembusan */}
-          <div className="sk-tembusan mt-10">
-            <p className="m-0">Tembusan :</p>
-            <p className="m-0">1.&nbsp; Kepala Biro Umum Kementerian Kehutanan</p>
-            <p className="m-0">2.&nbsp; Sekretaris Direktorat Jenderal KSDAE</p>
-          </div>
-        </div>
-      </article>
-
-      {/* ── HALAMAN 3 DST: Lampiran — Tabel Daftar Penghentian ── */}
-      {attachmentPages.map((page, index) => (
-        <article
-          className="sk-page sk-attachment-document mx-auto max-w-[210mm] bg-white px-24 py-12 text-black shadow-xl ring-1 ring-zinc-200"
-          style={pageStyle}
-          key={`sk-attachment-${index}`}
-        >
-          <div className="sk-body mx-auto w-[166mm]">
-            {page.includeMeta && renderAttachmentMetaTitle()}
-            {renderAttachmentTable(page.assets, page.startIndex, page.showColumnNumbers, page.includeTotal)}
-            {page.includeSignature && renderAttachmentSignature()}
-          </div>
-        </article>
-      ))}
-
     </div>
   );
 }
